test sample with 1 elements
